+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added baselime mark command
 
+### Removed
+- Removed unnecessary options in baselime console command
+
 ## [0.0.72] 2023-06-30
 ### Fixed
 - Fixed data endpoint