+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Removed
+- Removed _firstSeen and _lastSeen from the query results
+
 
 ## [0.0.73] 2023-07-07
 ### Added